refactor(#1696): split reader/mod.rs backend resolution to keep the file-size ratchet green
F4's fix added 10 lines to `cqlite-core/src/storage/sstable/reader/mod.rs`, which
is 1721/800 — over the source target, so the `file-size` ratchet FAILs any growth
and `CQLITE_ALLOW_FILE_GROWTH=1` is not an option here. The sanctioned response
is a split (epic #1116), so the disk-access backend resolution moves to a sibling
`reader/backend_resolve.rs`: the `CQLITE_*` env overrides, the pure
`parse_*`/`resolve_disk_access_mode` decisions, `system_memory_bytes`,
`direct_io_available` and `mmap_advice_for` — one cohesive responsibility
(config + env + file size -> backend + madvise), all of it already pure and
unit-tested.

A pure move: no behavior change, `pub(super)` visibility, mod.rs imports the six
names it calls (`mmap_advice_for` stays `#[cfg(unix)]`), and `reader::tests`
reaches the three parse helpers through the new module path. mod.rs 1721 -> 1560.
